handle directory notification with incremental transforms.
The Slicer transform is incremental, yet was not ready to get notification
of directory added, changed or removed.

(cherry picked from commit 3ba3cb71f01368209e0eda6536a81e49b47b58b3)

+++ build-system/gradle-core/src/main/java/com/android/build/gradle/internal/transforms/InstantRunSlicer.java
@@ -363,13 +363,22 @@ public class InstantRunSlicer extends Transform {
                         case ADDED:
                         case CHANGED:
                             Files.createParentDirs(outputFile);
-                            Files.copy(fileToProcess, outputFile);
+                            if (fileToProcess.isDirectory()) {
+                                FileUtils.copyDirectoryToDirectory(fileToProcess, outputFile);
+                            } else {
+                                Files.copy(fileToProcess, outputFile);
+                            }
                             break;
                         case REMOVED:
-                            if (!outputFile.delete()) {
-                                throw new TransformException(
-                                        String.format("Cannot delete file %1$s",
-                                                outputFile.getAbsolutePath()));
+                            if (outputFile.exists()) {
+                                if (outputFile.isDirectory()) {
+                                    FileUtils.deleteDirectoryContents(outputFile);
+                                }
+                                if (!outputFile.delete()) {
+                                    throw new TransformException(
+                                            String.format("Cannot delete file %1$s",
+                                                    outputFile.getAbsolutePath()));
+                                }
                             }
                             break;
                         default:
